Testing environment

+++ tests/test_10_attributes.py
@@ -8,14 +8,14 @@ test_object_id = 1
 
 
 def test_GetAttributeIdByName():
-    assert isinstance(rt.rtapi.GetAttributeIdByName('HW type'), long) == True
+    assert isinstance(rt.rtapi.GetAttributeIdByName('HW type'), int) == True
 
 def test_GetAttributeId():
-    assert isinstance(rt.rtapi.GetAttributeId('HW type'), long) == True
+    assert isinstance(rt.rtapi.GetAttributeId('HW type'), int) == True
 
 def test_GetDictionaryChapterId():
     chapter_id = rt.rtapi.GetDictionaryChapterId('Yes/No')
-    assert isinstance(chapter_id, long) == True
+    assert isinstance(chapter_id, int) == True
 
 def test_IntertDictionaryValue():
     chapter_id = rt.rtapi.GetDictionaryChapterId('server models')
@@ -24,7 +24,7 @@ def test_IntertDictionaryValue():
 def test_GetDictionaryIdByValue():
     global test_dict_value
     test_dict_value = rt.rtapi.GetDictionaryIdByValue('testmodel')
-    assert isinstance(test_dict_value, long) == True 
+    assert isinstance(test_dict_value, int) == True 
 
 def test_GetDictionaryValueById():
     assert rt.rtapi.GetDictionaryValueById(test_dict_value) == 'testmodel'
